Expanded UserController web mappings & related classes.
With WebSecurityConfiguration.java, the application can define the web mappings that are allowed to be queried. The mapping *has* to be mapped; otherwise, a `403 Forbidden` will be returned,

Added UserController.addUser and UserController.deleteUser; the former is "fully" featured and the latter is still in development but should follow in the former's footsteps.

Other files updated accordingly.